Move verifier from captain to consensus
Part of the great repo cleanup
