+++ ironfish/src/testUtilities/verifier.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { Verifier } from '../captain'
+import { Verifier } from '../consensus'
 import {
   IronfishBlockchain,
   IronfishNoteEncrypted,
